pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cognitive-core"
version = "0.7.0"
description = "Искусственный мультимодальный мозг — когнитивная система с памятью, восприятием и автономным циклом"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "cognitive-core contributors"},
]
keywords = ["ai", "cognitive", "brain", "multimodal", "memory"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.26.0",
    "psutil>=5.9.0",
    "jsonlines>=4.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
    "bandit>=1.7",
    "hypothesis>=6.112.0",
    "mutmut>=2.4.4",
]
nlp = [
    "pymorphy3>=1.0.0",
    "razdel>=0.5.0",
    "nltk>=3.8.0",
    "navec>=0.10.0",
    "sentence-transformers>=2.7.0",
]
vision = [
    "open-clip-torch>=2.24.0",
    "pillow>=10.0.0",
]
audio = [
    "openai-whisper>=20231117",
]
docs = [
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "rapidocr-onnxruntime>=1.2.3",
]
apidocs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.25.0",
]
encrypted = [
    "sqlcipher3>=0.5.0",
]
openai = [
    "openai>=1.0",
]
anthropic = [
    "anthropic>=0.20",
]
all = [
    "cognitive-core[dev,nlp,vision,audio,docs,apidocs,encrypted,openai,anthropic]",
]

[project.scripts]
cognitive-core = "brain.cli:cli_entry"

[tool.setuptools.packages.find]
include = ["brain*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = false
disallow_untyped_defs = false

[tool.bandit]
exclude_dirs = ["tests", "examples", "docs"]
skips = ["B101", "B110"]  # B101: assert в контрактах; B110: try/except/pass в cleanup/teardown

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "SIM", "C4", "RET", "PIE"]
ignore = [
    "RET504",   # unnecessary-assign — стилистическое предпочтение
    "SIM102",   # collapsible-if — ухудшает читаемость
    "SIM103",   # needless-bool — ложные срабатывания
    "SIM105",   # suppressible-exception — contextlib.suppress не всегда лучше
    "SIM108",   # if-else-block-instead-of-if-exp — ухудшает читаемость
    "SIM110",   # reimplemented-builtin — ложные срабатывания
    "SIM115",   # open-file-with-context-handler — не всегда применимо
    "SIM905",   # split-static-string — микрооптимизация
    "B905",     # zip-without-explicit-strict — Python 3.10+ strict kwarg
]
